move treetablenamecell
