+++ src/frontend/src/views/Data/components/TreesView/components/TreesTable/style.ts
@@ -1,0 +1,7 @@
+import styled from "@emotion/styled";
+// TODO-TR: import
+import { SortableHeader } from "../../../SamplesView/components/SamplesTable/components/SortableHeader";
+
+export const StyledSortableHeader = styled(SortableHeader)`
+  width: "auto";
+`;
